requirements-test.txt
app_version
django-appconf
pylru
mock
django-override-settings
factory_boy
